Upgraded TreadBot control system to 2022.1.1.
While I was at it, I added support in Robot.java for tandem use of
joysticks *and* the keyboard (via NetworkTables.)  The XBoxController
will have to wait; for one, we can't find it, and secondly,
controller.isConnected() reports true when the joysticks are connected!
To overcome that, we need to test for specific controller names, which
is both silly and something we have done in the past.  So we'll it again
if we have to.